+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva Backseat Driver
 
 ## [Unreleased]
 
+- [Update tool ids to qualify them as Clojure](https://github.com/BetterThanTomorrow/calva-backseat-driver/issues/34)
+
 ## [v0.0.19] - 2025-09-11
 
 - Fix: [The structural insert/replace top-level forms tools do not prevent editing non-top-level forms](https://github.com/BetterThanTomorrow/calva-backseat-driver/issues/32)